Adds static factory methods vs builders and reformating
This is a big commit which I am not proud of.
This commit has added static factory methods for the
Implementation classes vs the builders. Reasoning being
the builder classes were a fun exercise but the static factory
method makes more sense at this time.

Updates the formatting for the classes which was a big change to
every class.

Updates most of the tests to use the new static factory methods
and also updates the Serialization test class with the round trip
method.

whew.